requirements.txt
torch==2.6.0
tqdm==4.67.1
Pillow==10.4.0
bs4==0.0.2
scikit-learn==1.6.0
torchvision==0.21.0
wandb==0.19.1
lxml==5.3.0
accelerate==1.1.0
kaggle==1.6.17
gradio==5.14.0